Changed some logic. added nextDiscount "logic". created some products in main.

+++ src/main/java/Discount.java
@@ -1,6 +1,6 @@
 public interface Discount {
 
-    public int apply(Product product);
+    public double apply(Product product);
 
     public String getDescription(Product product);
